# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ExamCrack AI Coach API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the current user's student profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"

  # ── Dashboard ────────────────────────────────────────────────────────────────
...
  /syllabus/topics/{id}/mastery:
    patch:
      operationId: updateTopicMastery
      tags: [syllabus]
      summary: Update user mastery score for a topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasteryUpdate"
      responses:
        "200":
          description: Updated topic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyllabusTopic"
...
  /planner/tasks/{id}/complete:
    patch:
      operationId: completeTask
      tags: [planner]
      summary: Mark a task as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCompleteInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyTask"
  /planner/study-log:
    post:
      operationId: logStudySession
      tags: [planner]
      summary: Log study hours for today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyLogInput"
      responses:
        "201":
          description: Session logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySession"

  # ── Progress ──────────────────────────────────────────────────────────────────
...
  /ai/explain:
    post:
      operationId: explainTopic
      tags: [ai]
      summary: Get AI explanation for a topic (mock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplainTopicInput"
      responses:
        "200":
          description: AI explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIExplanation"
  /ai/question:
    post:
      operationId: generateQuestion
      tags: [ai]
      summary: Generate a practice question (mock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "200":
          description: Practice question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeQuestion"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Onboarding ───────────────────────────────────────────────────────────
...
    StudentProfile:
      type: object
      required:
        - id
        - clerkUserId
        - name
        - category
        - qualification
        - targetExam
        - dailyStudyHours
        - previousAttempts
        - strengthAreas
        - weakAreas
        - learningStyle
        - language
        - successProbabilityBaseline
        - createdAt
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        name:
          type: string
        category:
          type: string
        qualification:
          type: string
        targetExam:
          type: string
        dailyStudyHours:
          type: number
        previousAttempts:
          type: integer
        strengthAreas:
          type: array
          items:
            type: string
        weakAreas:
          type: array
          items:
            type: string
        learningStyle:
          type: string
        language:
          type: string
        successProbabilityBaseline:
          type: number
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        dailyStudyHours:
          type: number
        learningStyle:
          type: string
        language:
          type: string
        strengthAreas:
          type: array
          items:
            type: string
        weakAreas:
          type: array
          items:
            type: string

    # ── Dashboard ────────────────────────────────────────────────────────────
...
    SyllabusTopic:
      type: object
      required:
        - id
        - subject
        - topicName
        - weightage
        - pyqFrequency
        - difficultyScore
        - masteryScore
        - priorityScore
        - status
      properties:
        id:
          type: integer
        subject:
          type: string
        topicName:
          type: string
        weightage:
          type: number
          description: Exam weightage (0-100)
        pyqFrequency:
          type: integer
          description: Number of times appeared in PYQs
        difficultyScore:
          type: number
          description: Difficulty 1-10
        masteryScore:
          type: number
          description: User mastery 0-100
        priorityScore:
          type: number
          description: Calculated priority score
        status:
          type: string
          enum: [not_started, in_progress, mastered, revision_needed]
    MasteryUpdate:
      type: object
      required:
        - masteryScore
      properties:
        masteryScore:
          type: number
...
    StudyTask:
      type: object
      required:
        - id
        - title
        - subject
        - topicName
        - durationMinutes
        - taskType
        - completed
        - scheduledDate
      properties:
        id:
          type: integer
        title:
          type: string
        subject:
          type: string
        topicName:
          type: string
        durationMinutes:
          type: integer
        taskType:
          type: string
          enum: [study, practice, revision, mock_test]
        completed:
          type: boolean
        scheduledDate:
          type: string
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
    TaskCompleteInput:
      type: object
      required:
        - completed
      properties:
        completed:
          type: boolean
    StudyLogInput:
      type: object
      required:
        - hours
        - date
      properties:
        hours:
          type: number
        date:
          type: string
          format: date
        notes:
          type: string
    StudySession:
      type: object
      required:
        - id
        - hours
        - date
        - createdAt
      properties:
        id:
          type: integer
        hours:
          type: number
        date:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── Progress ─────────────────────────────────────────────────────────────
...
    ExplainTopicInput:
      type: object
      required:
        - topicId
        - mode
      properties:
        topicId:
          type: integer
        mode:
          type: string
          enum: [beginner, intermediate, advanced]
        learningStyle:
          type: string
          enum: [story, visual, exam_oriented, quick_revision]
        language:
          type: string
          enum: [english, hindi, hinglish]
    AIExplanation:
      type: object
      required:
        - topicName
        - explanation
        - keyPoints
        - examTips
        - mode
      properties:
        topicName:
          type: string
        explanation:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        examTips:
          type: array
          items:
            type: string
        mode:
          type: string
        language:
          type: string
    QuestionInput:
      type: object
      required:
        - topicId
        - difficulty
      properties:
        topicId:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
    PracticeQuestion:
      type: object
      required:
        - id
        - topicName
        - question
        - options
        - correctAnswer
        - explanation
        - difficulty
      properties:
        id:
          type: integer
        topicName:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
        explanation:
          type: string
        difficulty:
          type: string
...