# lib/api-zod/src/generated/types/explainTopicInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ExamCrack AI Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { ExplainTopicInputLanguage } from './explainTopicInputLanguage';
import type { ExplainTopicInputLearningStyle } from './explainTopicInputLearningStyle';
import type { ExplainTopicInputMode } from './explainTopicInputMode';

export interface ExplainTopicInput {
  topicId: number;
  mode: ExplainTopicInputMode;
  learningStyle?: ExplainTopicInputLearningStyle;
  language?: ExplainTopicInputLanguage;
}