# lib/api-zod/src/generated/types/practiceQuestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ExamCrack AI Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeQuestion {
  id: number;
  topicName: string;
  question: string;
  options: string[];
  correctAnswer: number;
  explanation: string;
  difficulty: string;
}