# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ExamCrack AI Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface StudySession {
  id: number;
  hours: number;
  date: Date;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}