# lib/api-zod/src/generated/types/studyTask.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ExamCrack AI Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { StudyTaskTaskType } from './studyTaskTaskType';

export interface StudyTask {
  id: number;
  title: string;
  subject: string;
  topicName: string;
  durationMinutes: number;
  taskType: StudyTaskTaskType;
  completed: boolean;
  scheduledDate: Date;
  /** @nullable */
  completedAt?: Date | null;
}